DON'T PANIC
When fetching the job, instead of aborting the whole thing on a
failure (such as 404), print a friendly message to STDERR, ignore the
job and keep going.

Some jobs get skipped completely[1] which resulted in a crash. By
continuing, we are able to collect all the other jobs while still
letting the user know something's wrong.

By printing it to the standard error, this won't pollute the XML
that's optionally passed to a clipboard command or some other
reporting tool.

[1]: https://storage.googleapis.com/origin-ci-test/logs/release-openshift-ocp-installer-e2e-openstack-serial-4.3/827/started.json